定时任务博客浏览量完成

+++ src/main/java/com/love/outofmemory/service/IBlogService.java
@@ -56,5 +56,5 @@ public interface IBlogService {
 
     Integer getTotalcountbyclass(Integer id);
 
-    int saveredisViews(Map<Object, Object> viewmap);
+    Long saveredisViews(Map<Object, Object> viewmap);
 }
